Yuka entiti separated from main
